feat : 메인 장바구니 수정

+++ src/main/java/store/novabook/front/api/book/dto/response/GetBookSearchResponse.java
@@ -14,7 +14,7 @@ public record GetBookSearchResponse(
 	String image,
 	Long price,
 	Long discountPrice,
-	Integer score,
+	Double score,
 	Boolean isPackaged
 ) {
 }
